Implemented and Tested Scanner
Calc Scan looks at each token and assigns it a type.

The types are enumerated values.

First the scanner checks to see if the token is a reserved word or operation by looking in the Token_Type map.

If its not in their then it checks to see if its an ID or Number by running the isID and isNum function. If neither function is true then it is assigned an Undefined type.

Undefined tokens cause syntax errors.