factorTail method, debugging, rewrote parse method
factorTail is pretty much the same as the pseudo code, I had to create
2 conditionals (1 for mult_op and 1 for div_op) but no big deal

I fixed most of the syntax errors we had and type is a private
attribute of token so I called the accessor method where need be. Also
I changed the iterator to a token iterator so we don’t have to cast it
each time getNextToken is called.

parse does not need to iterate through each element, instead it calls
expression and then checks to see that no more elements are left.